Refactor loot handling
